+++ src/components/shared/index.ts
@@ -1,0 +1,8 @@
+export { DataTable } from "./DataTable";
+export { FacetFilter } from "./FacetFilter";
+export { FormModal } from "./FormModal";
+export { PageHeader } from "./PageHeader";
+export { StatsCard } from "./StatsCard";
+export { WorkspaceTabs } from "./WorkspaceTabs";
+export { AppHeader } from "./AppHeader";
+export { AppNavbar } from "./AppNavbar";